# io/rest/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    ZoneCreate:
        properties:
            capacity:
                format: int64
                type: integer
                x-go-name: Capacity
            enabled:
                type: boolean
                x-go-name: Enabled
            id:
                format: int64
                type: integer
                x-go-name: ID
            parking_id:
                format: int64
                type: integer
                x-go-name: ParkingID
            remained_capacity:
                format: int64
                type: integer
                x-go-name: RemainedCapacity
        type: object
        x-go-package: github.com/farazff/IoT-parking/io/rest/sw
    ZoneCreator:
        properties:
            capacity:
                type: string
                x-go-name: Capacity
            enabled:
                type: string
                x-go-name: Enabled
            remained_capacity:
                type: string
                x-go-name: RemainedCapacity
        required:
            - capacity
            - enabled
            - remained_capacity
        type: object
        x-go-package: github.com/farazff/IoT-parking/io/rest/sw
    ZoneGet:
        properties:
            enabled:
                type: boolean
                x-go-name: Enabled
            first_name:
                type: string
                x-go-name: FirstName
            id:
                format: int64
                type: integer
                x-go-name: ID
            last_name:
                type: string
                x-go-name: LastName
            parking_id:
                format: int64
                type: integer
                x-go-name: ParkingID
            phone:
                type: string
                x-go-name: Phone
        type: object
        x-go-package: github.com/farazff/IoT-parking/io/rest/sw
...
host: localhost:7676
info:
    contact:
        email: f.farangizadeh@gmail.com
        name: Faraz Farangizadeh
    description: Document of Parking-IoT api
    title: Parking IoT Swagger
    version: 1.0.0
paths:
...
    /v1/zone:
        post:
            operationId: createZone
            parameters:
                - description: api key for APIs
                  in: header
                  name: api-key
                  required: true
                  type: string
                  x-go-name: ApiKey
                - description: Session token
                  in: header
                  name: session_token
                  required: true
                  type: string
                  x-go-name: SessionToken
                - in: body
                  name: ZoneCreator
                  required: true
                  schema:
                    $ref: '#/definitions/ZoneCreator'
            responses:
                "201":
                    $ref: '#/responses/ZoneCreateRes'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to create zone
            tags:
                - Parking_Admin
    /v1/zone/{id}:
        delete:
            operationId: deleteZone
            parameters:
                - description: api key for APIs
                  in: header
                  name: api-key
                  required: true
                  type: string
                  x-go-name: ApiKey
                - description: Session token
                  in: header
                  name: session_token
                  required: true
                  type: string
                  x-go-name: SessionToken
                - format: int64
                  in: path
                  name: ID
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/ErrorMessage'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "404":
                    $ref: '#/responses/ErrorMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to delete a zone by ID
            tags:
                - Parking_Admin
        get:
            operationId: getZone
            parameters:
                - description: api key for APIs
                  in: header
                  name: api-key
                  required: true
                  type: string
                  x-go-name: ApiKey
                - description: Session token
                  in: header
                  name: session_token
                  required: true
                  type: string
                  x-go-name: SessionToken
                - format: int64
                  in: path
                  name: ID
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/ZoneGetRes'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "404":
                    $ref: '#/responses/ErrorMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to get a single zone by ID
            tags:
                - Parking_Admin
        put:
            operationId: updateZone
            responses:
                "201":
                    $ref: '#/responses/ZoneUpdateRes'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "404":
                    $ref: '#/responses/ErrorMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to update a zone
            tags:
                - Parking_Admin
    /v1/zones:
        get:
            operationId: getZones
            parameters:
                - description: api key for APIs
                  in: header
                  name: api-key
                  required: true
                  type: string
                  x-go-name: ApiKey
                - description: Session token
                  in: header
                  name: session_token
                  required: true
                  type: string
                  x-go-name: SessionToken
            responses:
                "200":
                    $ref: '#/responses/ZonesGetRes'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to get all zones
            tags:
                - Parking_Admin
produces:
    - application/json
responses:
    ErrorMessage:
        description: ""
        headers:
            session_token:
                description: 'in: cookie'
                type: string
        schema:
            properties:
                message:
                    description: Error message
                    type: string
                    x-go-name: Message
            type: object
    ErrorUnauthorizedMessage:
        description: ""
        schema:
            properties:
                message:
                    description: Error message
                    type: string
                    x-go-name: Message
            type: object
...
    ZoneCreateRes:
        description: ""
        headers:
            session_token:
                description: 'in: cookie'
                type: string
        schema:
            properties:
                zone:
                    $ref: '#/definitions/ZoneCreate'
            type: object
    ZonesGetRes:
        description: ""
        headers:
            session_token:
                description: 'in: cookie'
                type: string
        schema:
            properties:
                zones:
                    description: parking res
                    items:
                        $ref: '#/definitions/ZoneGet'
                    type: array
                    x-go-name: ZonesGet
            type: object
schemes:
    - http
swagger: "2.0"